requirements.txt
pytest==4.4.0
automata_lib==3.1.0.post1
attrs==19.1.0
